requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
pypdf
numpy
